Moved to plugin_utils.
